Add files via upload
Sign language recognition is a problem that has been addressed in research for years. However, we are still far from finding a complete solution available in our society. The world is hardly live without communication, no matter whether it is in the form of texture, voice or visual expression. This project has been aimed to overcome the challenges which has been faced years for having seamless communication between normal and hearing-impaired person so that there has a common platform of understanding and no major difficulties to pass on messages for deaf people. We have used python as a programming language for this project. We used LSTM as Algorithm as well as Open CV, NumPy used as library. Tenserflow is a framework used for this project. This project is based on CNN (Conventional Neural Network) which is type of deep learning neural architecture which commonly use in computer vision which facilitates user to increase accuracy of interpreting visual data